Update index.md
mentioning the manifest pane and having an image of it was irrelevant and distracting.  We just want to know where to head, and that leads the reader to believe that they should be in the manifest pane before they continue reading.  It's like telling someone "You see a fork in the road, and you will want to head not left".