Update CSP link to be a little better (#1135)
Before this change, the mv3 manifest page had an inaccurate summary of what an extension's CSP should look like (a string instead of an object). We do not currently have a page describing valid CSP values in Manifest V3, so the link has been temporarily updated to the next best thing.