Update declarativeNetRequest docs for recent static ruleset changes
Recently we increased the number of static rulesets that an extension
can include, whilst limiting the number of those rulesets that an
extension can enable at one time. We also started to defer static
ruleset indexing for packed extensions until the ruleset is enabled
and with that started to ignore invalid static rules for packed
extensions.

See https://crbug.com/1209218 and https://crbug.com/1231280