Update the UA Reduction Origin Trial Blog Post
In M96, we are adding support for third-party embeds to register for the UA Reduction origin trial without requiring the top-level site in which it is embedded to be enrolled in the Origin Trial. This CL adds a section to the UA Reduction Blog Post to describe the origin trial for third-party requests.