Updated usage of executeScript prop name from function to func
While following the existing getting started article, it occurred that the updation of the background color of a tab use case is not working. Digging up a little in detail, found from https://developer.chrome.com/docs/extensions/reference/scripting/#runtime-functions. The proposed prop name is `func`. Whereas in the snippet of this getting started page we are maintaining prop name `function`.