Clarify wording for "3rd party matching" use case
The original wording "If you work with embedded content..." might be misleading, as almost every website is "working with embedded content", like e.g. "analytics scripts", "payment scripts", "widgets", etc. Developers might think that they need to follow these steps if they want to test this "embedded content" on their website, which is not the intended use case.

The new wording should clarify that this section is meant for "providers of embedded content".