Fix "get current window" example (#2770)
The `currentWindow` property in the queryOptions object matches the context
from which the call was executed. As such, using the example as previously
written in an extension tab or iframe would return information about the window
where this code was called, not the user's currently focused tab (which may be
in a different window).