Updated Webview User Agent to a later version. (#1654)
* Updated Webview User Agent to a later version.

The Webview UA was written up to Lollipop and Above in the article, which was leading to confusion for people who use App creating libraries like MIT App Inventor/Kodular/Appybuilder/Thunkable/Niotron or similar sites. I updated it to Android 10 and Above, so now it will be easier to put the more updated WebView UA string in their app's WebView component since customizing the WebView UA string may seem troublesome for many. SM-G960F is used as it is a more common device at the current time.

* Update site/en/docs/multidevice/user-agent/index.md

Co-authored-by: Rachel Andrew <me@rachelandrew.co.uk>

* Update site/en/docs/multidevice/user-agent/index.md

Co-authored-by: Rachel Andrew <me@rachelandrew.co.uk>

Co-authored-by: Rachel Andrew <me@rachelandrew.co.uk>