Playlist element #3959
I've created the base HTML and CSS for this element, and added it to the index for Workbox ready for the required 11ty magic to be applied to it.

The Figma had no mobile designs that I could find, will update the CSS upon receipt of the mobile design.

Ticket: https://github.com/GoogleChrome/developer.chrome.com/issues/3959