Make the capitalization of the #ssp anchor consistent (#4803)
The anchor to the glossary entry for "Supply-side platform, Sell-side platform" was #SSP, which didn't work for links that used the all-lower-case #ssp. On the glossary page itself, there were two links to #SSP and one to #ssp. From all other pages on the GoogleChrome/developer.chrome.com repo, the link was always to #ssp. (https://github.com/GoogleChrome/developer.chrome.com/search?q=%22%23SSP%22). The anchor for the glossary entry for "Demand-side platform (DSP)" is the all-lower-case #dsp. As such, changing SSP's anchor to all-lower-case as well.