Add details around third-party domain full UA access.
Additional details to help clarify what happens with the intersection of 1P and 3P sites participating in reverse origin trials. By suggesting they allow full ua access to all subresources (3ps), to help direct 3ps registering to click the 'third-party matching' box and to also try to avoid a situation where 3p's register for the ROT, the 1P also registers for the ROT and then 1P neglects to allow 3Ps to have the full UA (they need to do this or 3ps lose access to the full UA regardless of their own ROT registration.