Explain that Native Messaging is not available from content scripts (#5591)
* Explain that Native Messaging is not available from content scripts

https://chromium.googlesource.com/chromium/src/+/main/chrome/common/extensions/api/_features.md#Feature-Contexts

https://source.chromium.org/chromium/chromium/src/+/main:extensions/common/api/_api_features.json

  "runtime.connectNative": {
    "dependencies": ["permission:nativeMessaging"],
    "contexts": ["blessed_extension"]
  },

  "runtime.sendNativeMessage": {
    "dependencies": ["permission:nativeMessaging"],
    "contexts": ["blessed_extension"]
  },


The blessed_extension context refers to a JavaScript context running from an extension process. These are typically the most secure JavaScript contexts, as it reduces the likelihood that a compromised web page renderer will have access to secure APIs.

The content_script context refers to a JavaScript context for an extension content script. Since content scripts share a process with (and run on the same content as) web pages, these are considered very insecure contexts. Very few features should be exposed to these contexts.

* fix wording as requested