Fixing links in match_patterns.md
I've updated the syntax on the "grant access" link in line 36, but none of the internal links are working for me. Wondering if that's a permissions issue for me as a TVC.
I can make the links work using the full path (/site/en/docs...), but specifying a language in a file path is bad for localization.